udpate home page

+++ database/migrations/2021_12_31_120832_create_general_settings_table.php
@@ -20,7 +20,7 @@ class CreateGeneralSettingsTable extends Migration
       $table->string('app_logo_black')->nullable();
       $table->decimal('shipping_cost')->default(0);
       $table->string('shipping_days')->default(0);
-      $table->integer('tax')->default(0);
+      $table->decimal('tax')->default(0);
       $table->string('app_phone')->nullable();
       $table->string('app_email')->nullable();
       $table->string('app_address')->nullable();
